core/pyproject.toml
[project]
name = "oas-core"
version = "0.1.0"
description = "Opensens Agent Swarm — shared framework (swarm, middleware, protocols, adapters)"
requires-python = ">=3.11"

dependencies = [
    "pydantic>=2.0",
    "httpx>=0.27",
    "PyYAML>=6.0",
    "watchdog>=4.0",
]

[project.optional-dependencies]
swarm = [
    "langgraph>=1.0",
    "langgraph-swarm>=0.1.0",
    "langchain-anthropic>=0.3",
]
memory = [
    "openviking>=0.1.0",
]
redis = [
    "redis[hiredis]>=5.0",
]
wiki = [
    "lancedb>=0.6",
    "mlx-embeddings>=0.1",
]
all = [
    "oas-core[swarm,memory,redis,wiki]",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100
